+++ src/views/Settings/About.Settings.vue
@@ -8,40 +8,27 @@
         <a href="https://raw.githubusercontent.com/e6Hub/app/master/LICENSE" v-text="$t('settings.about.aboutLicense')"/>
       </i18n>
     </div>
-    <p>{{ require("../../../package.json").version }}</p>
     <div id="e6h__about_actions" class="mt-4">
       <btn
-        v-if="updateStatus == 'pendingRestart'"
-        @click.native="restartForUpdate"
-        role="safe"
-        v-text="$t('settings.about.restartToUpdate')"
-      />
-      <btn
-        v-else-if="updateStatus == 'downloading'"
-        role="busy"
-        v-text="$t('settings.about.downloadingUpdate')"
-      />
-      <btn
-        v-else-if="updateStatus == 'uptodate'"
-        @click.native="checkForUpdate"
-        v-text="$t('settings.about.checkForUpdate')"
-      />
-      <btn
-        v-else-if="updateStatus == 'checking'"
-        role="disabledBusy"
-        v-text="$t('settings.about.checkingForUpdate')"
-      />
-      <btn
-        v-else-if="updateStatus == 'dev'"
-        role="disabled"
-        v-text="$t('settings.about.devBuild')"
+        v-if="currentUpdaterStatus"
+        @click.native="currentUpdaterAction"
+        :role="currentUpdaterRoleButton"
+        v-text="currentUpdaterStatus"
+        class="mr-4"
       />
       <btn
         @click.native="resetSets"
         role="warn"
-        class="ml-4"
         v-text="$t('settings.about.resetSettings')"
       />
+    </div>
+    <div id="e6h__about_version" class="mt-8 text-gray-6 text-center select-text">
+      <i18n path="settings.about.versionDetail" tag="p">
+        <span v-text="require('../../../package.json').version"/>
+        <span v-text="platform.type"/>
+        <span v-text="platform.release"/>
+        <span v-text="platform.arch"/>
+      </i18n>
     </div>
   </SettingView>
 </template>
@@ -51,28 +38,78 @@ import SettingView from "@/components/SettingView.vue";
 import Btn from "@/components/Button.vue";
 import { ipcRenderer } from "electron";
 import { mapGetters, mapActions } from "vuex";
+import os from 'os';
 
 export default {
   name: "AboutSettings",
   components: { SettingView, Btn },
-  computed: {
-    ...mapGetters("updater", ["updateStatus"]),
+  data() {
+    return {
+      currentUpdaterAction: null,
+      currentUpdaterStatus: this.updateStatus,
+      currentUpdaterRoleButton: 'disabledBusy',
+      platform: {
+        type: os.type(),
+        arch: os.arch(),
+        release: os.release()
+      },
+
+    }
+  },
+  computed: mapGetters("updater", ["updateStatus"]),
+  watch: {
+    updateStatus() {
+      this.onUpdateStatus();
+    }
   },
   methods: {
     ...mapActions("settings", ["resetSettings"]),
+    onUpdateStatus() {
+      switch (this.updateStatus) {
+        case 'pendingRestart':
+          this.currentUpdaterStatus = this.$t('settings.about.restartToUpdate');
+          this.currentUpdaterAction = this.restartForUpdate;
+          this.currentUpdaterRoleButton = 'safe';
+          break;
+        case 'downloading':
+          this.currentUpdaterStatus = this.$t('settings.about.downloadingUpdate');
+          this.currentUpdaterAction = null;
+          this.currentUpdaterRoleButton = 'busy';
+          break;
+        case 'uptodate':
+          this.currentUpdaterStatus = this.$t('settings.about.checkForUpdate');
+          this.currentUpdaterAction = this.checkForUpdate;
+          this.currentUpdaterRoleButton = 'normal';
+          break;
+        case 'checking':
+          this.currentUpdaterStatus = this.$t('settings.about.checkingForUpdate');
+          this.currentUpdaterAction = null;
+          this.currentUpdaterRoleButton = 'disabledBusy';
+          break;
+        case 'dev':
+          this.currentUpdaterStatus = this.$t('settings.about.devBuild');
+          this.currentUpdaterAction = null;
+          this.currentUpdaterRoleButton = 'disabled';
+          break;
+        default:
+          this.currentUpdaterStatus = this.$t('settings.about.unsupportedBuild');
+          this.currentUpdaterAction = null;
+          this.currentUpdaterRoleButton = 'disabled';
+          break;
+      }
+    },
     restartForUpdate() {
       ipcRenderer.send("restartForUpdate");
     },
     checkForUpdate() {
       ipcRenderer.send("checkForUpdates");
     },
-    openExternal(e) {
-      e.preventDefault();
-      require("electron").shell.openExternal(e.target.getAttribute("href"));
-    },
     resetSets() {
       this.resetSettings();
     },
   },
+  created() {
+    this.onUpdateStatus();
+  }
 };
 </script>